+++ changelog.md
@@ -12,3 +12,10 @@
 **v0.1.2**
 - December 30, 2024
 - Revise the Landing Page Contents
+
+**v0.2.0**
+- January 2, 2025
+- Add Tabler Assets
+- Add Blank Page for Dashboard
+- Add Route files for route separation
+- Update Main web.php route